fix: rename fix test
